backend/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
flasgger==0.9.7b2
Flask==3.0.2
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
netifaces==0.11.0
packaging==24.0
pymongo==4.6.2
python-nmap==0.7.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
simplejson==3.19.2
six==1.16.0
Werkzeug==3.0.1
